docs(form): update stories to csf3 for Training Form  component
